keep the accessible buffer visible until `tab/escape` are used (#176054)
