+++ package.json
@@ -87,13 +87,13 @@
     "vscode-proxy-agent": "^0.12.0",
     "vscode-regexpp": "^3.1.0",
     "vscode-textmate": "9.0.0",
-    "xterm": "5.2.0-beta.29",
+    "xterm": "5.2.0-beta.30",
     "xterm-addon-canvas": "0.4.0-beta.7",
     "xterm-addon-search": "0.11.0",
     "xterm-addon-serialize": "0.9.0",
     "xterm-addon-unicode11": "0.5.0",
     "xterm-addon-webgl": "0.15.0-beta.7",
-    "xterm-headless": "5.2.0-beta.29",
+    "xterm-headless": "5.2.0-beta.30",
     "yauzl": "^2.9.2",
     "yazl": "^2.4.3"
   },
