+++ remote/web/package.json
@@ -11,7 +11,7 @@
 		"tas-client-umd": "0.1.6",
 		"vscode-oniguruma": "1.7.0",
 		"vscode-textmate": "9.0.0",
-		"xterm": "5.2.0-beta.29",
+		"xterm": "5.2.0-beta.30",
 		"xterm-addon-canvas": "0.4.0-beta.7",
 		"xterm-addon-search": "0.11.0",
 		"xterm-addon-unicode11": "0.5.0",
