+++ src/vs/workbench/contrib/terminal/browser/media/terminal.css
@@ -587,18 +587,28 @@
 	outline-color: var(--vscode-focusBorder);
 }
 
-/* Overrides for the a11y buffer as we're hosting monaco within it */
-.xterm .xterm-accessible-buffer {
-	padding: 0 !important;
-	overflow: initial !important;
-	overflow-x: initial !important;
-}
-.xterm .xterm-accessible-buffer:focus-within {
-	top: 0 !important;
-	left: 0 !important;
-	pointer-events: all !important;
-}
-.xterm > .xterm-accessible-buffer {
+.monaco-workbench .accessible-buffer {
+	position: absolute;
+	left: 0;
+	top: 0;
+	bottom: 0;
+	right: 0;
+	opacity: 0;
 	/* Reset cursor style as monaco controls it here */
 	cursor: default;
+	padding: 0;
+	overflow: initial;
+	overflow-x: initial;
+}
+
+.monaco-workbench .accessible-buffer div {
+	white-space: pre-wrap;
+}
+
+.monaco-workbench .accessible-buffer.focus-within,
+.monaco-workbench .accessible-buffer.active {
+	pointer-events: all;
+	opacity: 1;
+	z-index: 33;
+	background-color: var(--vscode-terminal-background, --vscode-panel-background);
 }
