+++ src/vs/workbench/contrib/terminal/browser/media/xterm.css
@@ -209,27 +209,6 @@
 	position: relative;
 }
 
-.xterm .xterm-accessible-buffer {
-	position: absolute;
-	left: 0;
-	top: 0;
-	bottom: 0;
-	right: 0;
-	padding: .5em;
-	opacity: 0;
-	overflow: scroll;
-	overflow-x: hidden;
-}
-
-.xterm .xterm-accessible-buffer div {
-	white-space: pre-wrap;
-}
-
-.xterm .xterm-accessible-buffer:focus-within {
-	opacity: 1;
-	z-index: 33;
-	top: -5px;
-	left: 5px;
-	pointer-events: none;
-	background-color: var(--vscode-terminal-background, --vscode-panel-background);
+.xterm.terminal.hide {
+	display: none;
 }
